% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_print.R
\name{default_print}
\alias{default_print}
\title{Use the default print method for the table class}
\usage{
default_print(x, ...)
}
\arguments{
\item{x}{the table to print}

\item{...}{args forwarded to the default print function}
}
\description{
This function will dispatch the default print method for the table, if any of
the table's classes have been registered for flipping. Where the table has
multiple classes registered, the priority is given to the order specified in
the registration config.
}
