% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Cycle through configured print methods}
\usage{
flip(last_value = .Last.value)
}
\arguments{
\item{last_value}{The value to be printed with the
next active print method cycled to by \code{flip()} defaults to \code{.Last.value}. Mainly used for testing.}
}
\description{
`print()`` methods are kept in the order supplied in \code{\link[=register_flips]{register_flips()}}, calling
this function cycles between them changing the active print method.
}
\details{
if the \code{.Last.value} hs a class in \code{printed_classes} supplied to
\verb{[register_flips()]}, it is reprinted using the current print method when this
function is called.
}
