% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_on_off.R
\name{flip_off}
\alias{flip_off}
\title{Disable cycling between print methods with flip()}
\usage{
flip_off()
}
\description{
This undoes the print method rerouting done by \code{\link[=flip_on]{flip_on()}}. Print methods are
returned to defaults for all configured classes. The configuration created by
\code{\link[=register_flips]{register_flips()}} remains and can be enabled again using \code{\link[=flip_on]{flip_on()}}.
}
