% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_on_off.R
\name{flip_on}
\alias{flip_on}
\title{Enable cycling between print methods with flip()}
\usage{
flip_on()
}
\description{
When a configuration has been created with \code{\link[=register_flips]{register_flips()}} this function
uses that config to reroute the print methods of the configured classes,
enabling printing with user-defined methods and cycling between them.
}
