% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_flips.R
\name{print_override}
\alias{print_override}
\title{Create an object class - package namespace pair for print override}
\usage{
print_override(class, pkg_namespace)
}
\arguments{
\item{class}{the class of the object to override the print for}

\item{pkg_namespace}{the package namespace where the print method can be found}
}
\description{
The combination of class and namespace is used to determine the name of the
print method that is to be rerouted (overidden). E.g. \code{class = "data.frame"}
and \code{pkg_namespace = "base"} leads to a call like
\code{utils::getFromNamespace(paste0("print.", class), pkg_namespace)} to get
\code{base::print.data.frame}.
}
