% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_flips.R
\name{register_flips}
\alias{register_flips}
\title{Register alternative print methods and classes to which they apply}
\usage{
register_flips(printer_fns, printed_classes)
}
\arguments{
\item{printer_fns}{a list of functions to be cycled between in order supplied.}

\item{printed_classes}{a lit of class / package namespace pairs, each created
with a call to \code{\link[=print_override]{print_override()}}. The print method to override is assumed to
be \verb{<pkg_namespace>:::<class>.print()}.}
}
\description{
This function configures the flipping between \code{print()} methods in the
current sesson, but does not enable it. \code{\link[=flip_on]{flip_on()}} and \code{\link[=flip_off]{flip_off()}} enable
and disable flipping between configured \code{print()} methods for configured
classes.
}
\section{Details}{

The order of \code{printed_classes} listed as \code{\link[=print_override]{print_override()}} calls is somewhat
important if \code{\link[=default_print]{default_print()}}\verb{is among}printer_fns\verb{. Where an object has multiple classes from those configured in }printed_classes\code{the}default_print()\verb{from the one ' appearing first in}printed_classes\verb{will be used. The exception to this is ' the class}data.frame` which is always used as
the last resort since most other ' table classes are a specialisation of it.
}

\examples{
\dontrun{
library(flippingtables)
register_flips(
		printer_fns = list(
			paint::paint,
			function(x) withr::with_options(list(width = 300), default_print(x)),
			function(x) withr::with_options(list(width = 40), default_print(x, n = 50))),
		printed_classes = list(
			print_override(class = "tbl", pkg_namespace = "pillar"),
			print_override(class = "data.frame", pkg_namespace = "base"),
			print_override(class = "data.table", pkg_namespace = "data.table")
		)
	)
flip_on() # now the configuration is live and can be used with flip()
}
}
